+++ tests/testthat/test-merge_mp3.R
@@ -1,0 +1,5 @@
+test_that("merge_mp3() runs without error", {
+  skip_if_not(interactive(), "Skipping in non-interactive mode")
+
+  expect_error(merge_mp3(), NA)
+})